Significantly changed the build script and included code
Build script no longer contains gateway/nameserver/proxy settings for
simplicity. Advanced users would be able to configure this on their own.
Updated the BCM2835 library used to version 1.50 from 1.44 and fixed the
hash. Updated dependency installation to coincide with a Raspbian Jessie
image rather than Wheezy. Removed minicom as it is not a requirement,
apache and sql may not be required either. Changed node-red install to
work with Jessie. Made the script clone the firmware repo instead of
extracting a zip and deleted said zip. Deleted all the included Python
SO code and i2c C code to remove duplicates across repositories. This
can all be found in the firmware repo which gets cloned by the script.
Changed compilation steps to match the new code locations in the cloned
directory. Added the Sensorian Interface and Sensorian Hub Client to be
cloned, dependency installed and compiled as well. Need to try to update
dependency versions if possible and remove PIL in favour of Pillow as
even the included firmware example programs memory leak when drawing to
the TFT LCD.