Changed cron to run Client instead of IP
IP, like most of the C examples, runs SPI_Close() at the end which
breaks the purely Python Sensorian Hub Client since it can't initialize
the SPI bus. Runs the Client on boot instead which shows the IP amongst
a bunch of other variables and can be closed from the display
gracefully.